requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.3
django-countries==7.6.1
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
phonenumber-field==0.0.1
phonenumbers==9.0.8
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
